# Conflicts:
#	public/assets/js/controller/customers.js
#	public/assets/js/controller/staffs-list.js
#	public/assets/js/controller/staffs.js

+++ public/assets/js/controller/staffs.js
@@ -25519,8 +25519,7 @@ module.exports = function(module) {
           currentValue: this.range ? [null, null] : null,
           userInput: null,
           popupVisible: !1,
-          position: {},
-          id: Date.now()
+          position: {}
         };
       },
       watch: {
@@ -25591,9 +25590,6 @@ module.exports = function(module) {
         },
         innerPopupStyle: function innerPopupStyle() {
           return D({}, this.position, this.popupStyle);
-        },
-        createIDByTime: function createIDByTime() {
-          return "datepicker" + this.id;
         }
       },
       mounted: function mounted() {
@@ -25729,25 +25725,6 @@ module.exports = function(module) {
           this.$emit("blur", e);
         },
         handleFocus: function handleFocus(e) {
-          var t = !0,
-              n = !1,
-              a = void 0;
-
-          try {
-            for (var i, r = document.querySelectorAll(".mx-datepicker-popup")[Symbol.iterator](); !(t = (i = r.next()).done); t = !0) {
-              var s = i.value;
-              s.id.toString() !== "datepicker" + this.id.toString() ? s.style.display = "none" : s.style.display = "block";
-            }
-          } catch (e) {
-            n = !0, a = e;
-          } finally {
-            try {
-              !t && r.return && r.return();
-            } finally {
-              if (n) throw a;
-            }
-          }
-
           this.popupVisible || (this.popupVisible = !0), this.$emit("focus", e);
         },
         handleKeydown: function handleKeydown(e) {
@@ -25894,9 +25871,6 @@ module.exports = function(module) {
         ref: "calendar",
         staticClass: "mx-datepicker-popup",
         style: e.innerPopupStyle,
-        attrs: {
-          id: e.createIDByTime
-        },
         on: {
           click: function click(e) {
             e.stopPropagation(), e.preventDefault();
@@ -26601,6 +26575,11 @@ var ctrStaffsVl = new Vue({
           that.backToList();
           return false;
         } else {
+          if (id == auth_staff_id) {
+            alert(messages["MSG06005"]);
+            return;
+          }
+
           if (confirm(messages["MSG06001"])) {
             staffs_service.deleteStaffs(id).then(function (response) {
               window.location.href = listRoute;
@@ -26770,7 +26749,7 @@ var CACHE = [],
 /*! no static exports found */
 /***/ (function(module, exports, __webpack_require__) {
 
-module.exports = __webpack_require__(/*! E:\MyProject\akita-erp\resources\assets\js\controller\staffs-vl.js */"./resources/assets/js/controller/staffs-vl.js");
+module.exports = __webpack_require__(/*! D:\petproject\akita-erp\resources\assets\js\controller\staffs-vl.js */"./resources/assets/js/controller/staffs-vl.js");
 
 
 /***/ })
